+++ .eslintrc.js
@@ -1,27 +1,36 @@
+'use strict';
+
 module.exports = {
-  "extends": ["ash-nazg/sauron", "plugin:testcafe/recommended"],
-  "parserOptions": {
-    "sourceType": "module"
+  extends: ['ash-nazg/sauron', 'plugin:testcafe/recommended'],
+  parserOptions: {
+    sourceType: 'module'
   },
   settings: {
     polyfills: [
       'URL'
     ]
   },
-  "env": {
-    "node": false,
-    "browser": true
+  env: {
+    node: false,
+    browser: true
   },
   overrides: [
     {
-      files: ["*.md"],
+      files: '.eslintrc.js',
+      extends: ['plugin:node/recommended-script'],
+      rules: {
+        'import/no-commonjs': 0
+      }
+    },
+    {
+      files: ['*.md'],
       rules: {
         'import/no-unresolved': 0,
         'no-console': 0
       }
     }
   ],
-  "rules": {
+  rules: {
     // Disable for now
     'require-unicode-regexp': 0,
     'max-len': 0